+++ src/ansibleguy-webui/aw/config/defaults.py
@@ -36,6 +36,7 @@ def _get_defaults_docker(var: str) -> any:
 # need to be referenced multiple times without import dependencies
 CONFIG_DEFAULTS = {
     'port': 8000,
+    'address': '127.0.0.1',
     'run_timeout': 3600,
     'path_run': '/tmp/ansible-webui',
     'path_play': getcwd(),